article_all_clear

+++ final-pjt-back/community/views.py
@@ -1,4 +1,3 @@
-
 from rest_framework.response import Response
 from rest_framework.decorators import api_view
 # Authentication Decorators
@@ -19,8 +18,8 @@ from .models import Article, Comment
 @permission_classes([IsAuthenticated]) # 인증된 경우만 권한을 준다.
 def article_list(request):
     if request.method == 'GET':
-        # articles = Article.objects.all()
-        articles = get_list_or_404(Article)
+        articles = Article.objects.all()
+        # articles = get_list_or_404(Article)
         serializer = ArticleListSerializer(articles, many=True)
         return Response(serializer.data)
 
@@ -36,14 +35,17 @@ def article_list(request):
 def article_detail(request, article_pk):
     # article = Article.objects.get(pk=article_pk)
     article = get_object_or_404(Article, pk=article_pk)
-
     if request.method == 'GET':
         serializer = ArticleSerializer(article)
         return Response(serializer.data)
     
     elif request.method == 'DELETE':
-        article.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        writer = article.user
+        if request.user == writer: #요청유저와 글쓴 유저가 같다면
+            serializer = ArticleSerializer(article)
+            article.delete()
+            return Response(serializer.data)
+        return Response(status=status.HTTP_403_FORBIDDEN)
 
     elif request.method == 'PUT':
         serializer = ArticleSerializer(article, data=request.data)
@@ -56,7 +58,7 @@ def article_detail(request, article_pk):
 def comment_list(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
     if request.method == 'GET':
-        comments = get_list_or_404(Comment)
+        comments = article.comment_set.all()
         serializer = CommentSerializer(comments, many=True)
         return Response(serializer.data)
     
@@ -80,14 +82,10 @@ def comment_detail(request, article_pk, comment_pk):
     elif request.method == 'DELETE':
         writer = comment.user
         if request.user == writer: #요청유저와 글쓴 유저가 같다면
+            serializer = CommentSerializer(comment)
             comment.delete()
-            return Response(status=status.HTTP_204_NO_CONTENT)
-        else:
-            bad = True
-            context = {
-                bad: 'bad'
-            }
-            return Response(context)
+            return Response(serializer.data)
+        return Response(status=status.HTTP_400_BAD_REQUEST)
 
     elif request.method == 'PUT':
         serializer = CommentSerializer(comment, data=request.data)
